requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.3
frozenlist==1.3.3
h11==0.14.0
idna==3.4
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
petl==1.7.12
pluggy==1.0.0
pyaml==21.10.1
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
uvicorn==0.20.0
yarl==1.8.1
